requirements.txt
Django==1.11.24
django-admin-sortable2==0.7.3
django-summernote==0.8.11.4
easy-thumbnails==2.6
Pillow==6.1.0
psycopg2==2.8.4
pytz==2019.2
six==1.12.0
transliterate==1.10.2
